+++ tests/api_resources/ai/conversations/test_messages.py
@@ -9,6 +9,7 @@ import pytest
 
 from telnyx import Telnyx, AsyncTelnyx
 from tests.utils import assert_matches_type
+from telnyx.pagination import SyncDefaultFlatPagination, AsyncDefaultFlatPagination
 from telnyx.types.ai.conversations import MessageListResponse
 
 base_url = os.environ.get("TEST_API_BASE_URL", "http://127.0.0.1:4010")
@@ -21,33 +22,43 @@ class TestMessages:
     @parametrize
     def test_method_list(self, client: Telnyx) -> None:
         message = client.ai.conversations.messages.list(
-            "conversation_id",
+            conversation_id="conversation_id",
         )
-        assert_matches_type(MessageListResponse, message, path=["response"])
+        assert_matches_type(SyncDefaultFlatPagination[MessageListResponse], message, path=["response"])
+
+    @pytest.mark.skip(reason="Mock server tests are disabled")
+    @parametrize
+    def test_method_list_with_all_params(self, client: Telnyx) -> None:
+        message = client.ai.conversations.messages.list(
+            conversation_id="conversation_id",
+            page_number=1,
+            page_size=1,
+        )
+        assert_matches_type(SyncDefaultFlatPagination[MessageListResponse], message, path=["response"])
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     def test_raw_response_list(self, client: Telnyx) -> None:
         response = client.ai.conversations.messages.with_raw_response.list(
-            "conversation_id",
+            conversation_id="conversation_id",
         )
 
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         message = response.parse()
-        assert_matches_type(MessageListResponse, message, path=["response"])
+        assert_matches_type(SyncDefaultFlatPagination[MessageListResponse], message, path=["response"])
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     def test_streaming_response_list(self, client: Telnyx) -> None:
         with client.ai.conversations.messages.with_streaming_response.list(
-            "conversation_id",
+            conversation_id="conversation_id",
         ) as response:
             assert not response.is_closed
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             message = response.parse()
-            assert_matches_type(MessageListResponse, message, path=["response"])
+            assert_matches_type(SyncDefaultFlatPagination[MessageListResponse], message, path=["response"])
 
         assert cast(Any, response.is_closed) is True
 
@@ -56,7 +67,7 @@ class TestMessages:
     def test_path_params_list(self, client: Telnyx) -> None:
         with pytest.raises(ValueError, match=r"Expected a non-empty value for `conversation_id` but received ''"):
             client.ai.conversations.messages.with_raw_response.list(
-                "",
+                conversation_id="",
             )
 
 
@@ -69,33 +80,43 @@ class TestAsyncMessages:
     @parametrize
     async def test_method_list(self, async_client: AsyncTelnyx) -> None:
         message = await async_client.ai.conversations.messages.list(
-            "conversation_id",
+            conversation_id="conversation_id",
         )
-        assert_matches_type(MessageListResponse, message, path=["response"])
+        assert_matches_type(AsyncDefaultFlatPagination[MessageListResponse], message, path=["response"])
+
+    @pytest.mark.skip(reason="Mock server tests are disabled")
+    @parametrize
+    async def test_method_list_with_all_params(self, async_client: AsyncTelnyx) -> None:
+        message = await async_client.ai.conversations.messages.list(
+            conversation_id="conversation_id",
+            page_number=1,
+            page_size=1,
+        )
+        assert_matches_type(AsyncDefaultFlatPagination[MessageListResponse], message, path=["response"])
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     async def test_raw_response_list(self, async_client: AsyncTelnyx) -> None:
         response = await async_client.ai.conversations.messages.with_raw_response.list(
-            "conversation_id",
+            conversation_id="conversation_id",
         )
 
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         message = await response.parse()
-        assert_matches_type(MessageListResponse, message, path=["response"])
+        assert_matches_type(AsyncDefaultFlatPagination[MessageListResponse], message, path=["response"])
 
     @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     async def test_streaming_response_list(self, async_client: AsyncTelnyx) -> None:
         async with async_client.ai.conversations.messages.with_streaming_response.list(
-            "conversation_id",
+            conversation_id="conversation_id",
         ) as response:
             assert not response.is_closed
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             message = await response.parse()
-            assert_matches_type(MessageListResponse, message, path=["response"])
+            assert_matches_type(AsyncDefaultFlatPagination[MessageListResponse], message, path=["response"])
 
         assert cast(Any, response.is_closed) is True
 
@@ -104,5 +125,5 @@ class TestAsyncMessages:
     async def test_path_params_list(self, async_client: AsyncTelnyx) -> None:
         with pytest.raises(ValueError, match=r"Expected a non-empty value for `conversation_id` but received ''"):
             await async_client.ai.conversations.messages.with_raw_response.list(
-                "",
+                conversation_id="",
             )